+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 1.  Add: Support for `Unity.Mathematics` package, you can now use `ShowInInspector`/`Button` to display/edit an `bool4` (and more) types in inspector ~~to play tic-tac-toc~~
 2.  Unity Fix: Unity forget to make a drawer for `Unity.Mathematics.half`. It's now handled by SaintsField
 3.  Fix: `ShowInInspector` no longer shows a dropdown field when no other options avaliable for the target general class/struct
+4.  Fix: `Button` with `IEnumerator` type didn't show errors when any error happens
 
 ## 5.14.4 ##